# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'New India Export / VIRASTRA API'
  version: 2.3.0
  description: "Production Express API for customer onboarding, OTP auth, KYC, cases, Razorpay payments, GST invoices, events, and admin ops.\n\n**Auth:** Email OTP only. Send `Authorization: Bearer <JWT>` for protected routes.\n\n**Money:** Amounts in INR rupees in API/DB; converted to paise only for Razorpay.\n\n**Envelope:** `{ success, data?, message? }` on success; `{ success: false, message, code?, details? }` on errors.\n\nLegacy vault/shipment routes under `/api/cases/...` exist only when `ENABLE_LEGACY_SHIPMENT=true`."
  contact:
    name: 'New India Export'
    email: support@newindiaexport.com
servers:
  -
    url: 'http://127.0.0.1:5001'
    description: Local
  -
    url: /
    description: 'Current host'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from OTP verify / login. Claims: sub, email, role, status.'
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 'Recommended for money / submit endpoints'
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: array
          items: {}
...
    KycProfile:
      type: object
      properties:
        legalName:
          type: string
        legalEntityName:
          type: string
        entityType:
          type: string
        incorporationDate:
          type: string
        turnover:
          type: string
        registeredAddress:
          type: string
        operatingCity:
          type: string
        signatoryName:
          type: string
        fullName:
          type: string
        designation:
          type: string
        panNumber:
          type: string
        pan:
          type: string
        aadhaarLast4:
          type: string
          pattern: '^\d{4}$'
      additionalProperties: true
...
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        name:
          type: string
        designation:
          type: string
        phone:
          type: string
...
    KycNeedsMore:
      type: object
      properties:
        reason:
          type: string
  responses:
    BadRequest:
      description: 'Validation / bad request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Missing or invalid JWT'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Insufficient role / ACL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/kyc/me:
    get:
      summary: 'Own KYC status / profile / uploads'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/kyc/me/profile:
    post:
      summary: 'Save KYC profile'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycProfile'
    put:
      summary: 'Save KYC profile (PUT)'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycProfile'
  /api/kyc/me/business:
    put:
      summary: 'Save KYC business fields (alias of profile)'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycProfile'
  /api/kyc/me/identity:
    put:
      summary: 'Save KYC identity fields (alias of profile)'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycProfile'
  '/api/kyc/me/documents/{docId}':
    post:
      summary: 'Upload KYC document'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: docId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
    delete:
      summary: 'Delete KYC document'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: docId
          in: path
          required: true
          schema:
            type: string
          description: null
  /api/kyc/me/submit:
    post:
      summary: 'Submit KYC for review'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          $ref: '#/components/parameters/IdempotencyKey'
  /api/kyc/queue:
    get:
      summary: 'Ops KYC review queue'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  '/api/kyc/{caseId}/approve':
    post:
      summary: 'Approve KYC'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
          description: 'Case id or legacy email'
  '/api/kyc/{caseId}/needs-more':
    post:
      summary: 'Request more KYC documents'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycNeedsMore'
  '/api/kyc/{caseId}/reject':
    post:
      summary: 'Reject KYC (alias of needs-more)'
      tags:
        - KYC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycNeedsMore'
...
  /api/me/profile:
    get:
      summary: 'User profile'
      tags:
        - Me
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
    put:
      summary: 'Update profile'
      tags:
        - Me
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
...